🎸 feat(blog): Use a consistent page layout (page-width / margin)
A hook has been created thats responsible for the layout of all pages.
It replaced the individual page stylings of each page.
On the CV page it broke the archivements section, which didn't properly break/wrap into
multiple lines because it now had too much space. Therefore it is now explicitly specified
to display a list view for a specific viewport.