🎸 feat(blog): Add skeleton-loading while fetching blog entries and content
In the process of adding loading animation the useFetch strategy in the blog component has also been changed.
It is now using the auto-managed stated because it's easier to read.
The skeleton component comes from a experimental mui library. But at the time of writing it was one of the
easier libraries out there.