⚙️  imp(dep): Update latest dependencies in the web project
Also remove some warnings in the Skills component where a type was misdeclared and the key was missing in the iterated elements.
The Link dependencies for MUI and react-router had to be reversed, because it wasn't working after the update anymore. Seems like something changed in the react-router Link that broke the composition capabilities via the component prop.