🐛 (blog): Reenable back navigation to blog overview
With the recent dependency upgrade to react-router v6 the back navigation was broken.
Now we also check if there is anything in the navigation history before returning to that state. This is a problem if the user openes the page via a direct link where there is no history present yet. The fallback will be the blog overview.