⚙️  imp(infra): keep api awake to improve response times
The current host (heroku) is constantly putting the api to sleep after an inactivity of > 30mins.
By calling the api every 30 minutes we make sure this never happens.
The workflow_dispatch trigger was added to enable a manual trigger via the github interface.
The api url was moved to a secret so it isn't hardcoded in this file but configurable via github. (So not really a secret)